+++ 3DEngine/IcoSphere.h
@@ -7,10 +7,10 @@ class IcoSphere : public Shape
 public:
 	IcoSphere(int recursion);
 private:
-	unsigned short AddVertex(const Vec3f& point);
-	unsigned short GetMiddlePoint(unsigned short indexA, unsigned short indexB);
+	unsigned int AddVertex(const Vec3f& point);
+	unsigned int GetMiddlePoint(unsigned int indexA, unsigned int indexB);
 
 private:
-	std::unordered_map<unsigned int, unsigned short> cache;
-	unsigned short index;
+	std::unordered_map<unsigned long, unsigned int> cache;
+	unsigned int index;
 };
